Move Board and Slot classes to boards package
